[add] main1.h파일 제작, 토성고리, 태양이미지 삽입 - main1파일의 선언부를 main1.h파일에서 관리 - 토성 고리 제작 - 태양 텍스쳐 삽입
