+++ SolarSystem12/freeglutProject3D/Debug/freeglutProject3D.log
@@ -4,17 +4,26 @@ c:\users\sejong\oss12\solarsystem12\freeglutproject3d\sun.h : warning C4819: The
 c:\users\sejong\oss12\solarsystem12\freeglutproject3d\axis.h : warning C4819: The file contains a character that cannot be represented in the current code page (949). Save the file in Unicode format to prevent data loss
 c:\users\sejong\oss12\solarsystem12\freeglutproject3d\earth.h : warning C4819: The file contains a character that cannot be represented in the current code page (949). Save the file in Unicode format to prevent data loss
 c:\users\sejong\oss12\solarsystem12\freeglutproject3d\disk.h : warning C4819: The file contains a character that cannot be represented in the current code page (949). Save the file in Unicode format to prevent data loss
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(241): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(252): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(263): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(275): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(286): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(297): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(308): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(521): warning C4244: '=': conversion from 'int' to 'float', possible loss of data
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(522): warning C4244: '=': conversion from 'int' to 'float', possible loss of data
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(523): warning C4244: '=': conversion from 'int' to 'float', possible loss of data
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(833): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
-c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(102): warning C4715: 'random': not all control paths return a value
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(72): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(83): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(94): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(106): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(117): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(128): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(139): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(29): warning C4189: 'win': local variable is initialized but not referenced
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(370): warning C4244: '=': conversion from 'int' to 'float', possible loss of data
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(371): warning C4244: '=': conversion from 'int' to 'float', possible loss of data
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(372): warning C4244: '=': conversion from 'int' to 'float', possible loss of data
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(672): warning C4305: 'argument': truncation from 'double' to 'GLfloat'
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(895): warning C4100: 'mY': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(895): warning C4100: 'mX': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(932): warning C4100: 'mY': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(932): warning C4100: 'mX': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(965): warning C4100: 'mY': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(965): warning C4100: 'mX': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(1070): warning C4100: 'mY': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.cpp(1070): warning C4100: 'mX': unreferenced formal parameter
+c:\users\sejong\oss12\solarsystem12\freeglutproject3d\main1.h(104): warning C4715: 'random': not all control paths return a value
   freeglutProject3D.vcxproj -> C:\Users\Sejong\OSS12\SolarSystem12\Debug\SolarSystem.exe
   freeglutProject3D.vcxproj -> C:\Users\Sejong\OSS12\SolarSystem12\Debug\SolarSystem.pdb (Full PDB)
