+++ SolarSystem12/freeglutProject3D/freeglutProject3D.vcxproj.filters
@@ -152,6 +152,9 @@
     <ClInclude Include="Background.h">
       <Filter>Header Files\Objects</Filter>
     </ClInclude>
+    <ClInclude Include="main1.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Image Include="..\..\..\..\..\..\Downloads\earth.bmp">
